Refactoring of hover logic
